\name{MKpower-package}
\alias{MKpower-package}
\alias{MKpower}
\docType{package}
\title{
Functions for Power Analysis.
}
\description{
Contains several functions for power analysis and sample size calculation.
}
\details{
\tabular{ll}{
Package: \tab MKpower\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2019-08-19\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats\cr
Suggests: \tab knitr, rmarkdown\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKpower)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
