\name{MKpower-package}
\alias{MKpower-package}
\alias{MKpower}
\docType{package}
\title{
Power Analysis and Sample Size Calculation.
}
\description{
Functions for power analysis and sample size calculation. Examples are Welch and Hsu t-Tests, Wilcoxon Rank Sum and Signed Rank Tests as well as power and sample size required for the evaluation of a diagnostic test(-system).
}
\details{
\tabular{ll}{
Package: \tab MKpower\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2019-11-08\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, matrixTests\cr
Suggests: \tab knitr, rmarkdown\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKpower)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
